requirements.txt
openai
python-dotenv
numpy
scikit-learn
Django
Pillow
